rota de tarefas

+++ controllers/user.js
@@ -3,10 +3,13 @@ const jwt = require('jsonwebtoken')
 const bcryptjs = require('bcryptjs');
 const { SECRET } = require('../utils/config');
 const userRouter = require('express').Router()
+const { extrairToken, extrairUser } = require('../utils/middleware')
 
+
+// Não sei se vai usar
+// Encontra usuário pelo id
 userRouter.get('/:id', async (request, response) => {
     const id = request.params.id
-    const user = await User.findById(id)
 
     if (!user) {
         return response.status(400).send({ error: 'Usuário não encontrado' })
@@ -15,26 +18,13 @@ userRouter.get('/:id', async (request, response) => {
     response.status(200).json(user)
 })
 
-userRouter.get('/', async (request, response) => {
-    const authorization = request.get('authorization')
-
-    if (!(authorization && authorization.startsWith('Bearer '))) {
-        return response.status(401).send({ error: 'autorização falhou' })
-    }
-
-    const token = authorization.replace('Bearer ', '')
-
-    const tokenDecodificado = jwt.verify(token, SECRET)
-
-    if (!tokenDecodificado) {
-        return response.status(401).send({ error: 'token inválido' })
-    }
-
-    const user = await User.findById(tokenDecodificado.id)
-
+// Encontra usuário com token
+userRouter.get('/', extrairToken, extrairUser, async (request, response) => {
+    const user = request.user
     response.status(200).json(user)
 })
 
+// Cria usuário
 userRouter.post('/', async (request, response) => {
     const { email, password, name } = request.body
 
@@ -49,9 +39,15 @@ userRouter.post('/', async (request, response) => {
     response.status(201).json(userSalvo)
 })
 
-userRouter.put('/:id', async (request, response) => {
+// Atualiza usuário com token 
+userRouter.put('/:id', extrairToken, async (request, response) => {
     const id = request.params.id
     const { email, name, password } = request.body
+    const tokenId = request.token.id
+
+    if (id != tokenId) {
+        return response.status(401).send({ error: 'Não autorizado' })
+    }
 
     const passwordHash = password && password.length >= 6
         ? await bcryptjs.hash(password, 10)
@@ -76,8 +72,14 @@ userRouter.put('/:id', async (request, response) => {
     response.status(200).json(userAtualizado)
 })
 
-userRouter.delete('/:id', async (request, response) => {
+// Deleta usuário com token 
+userRouter.delete('/:id', extrairToken, async (request, response) => {
     const id = request.params.id
+    const tokenId = request.token.id
+
+    if (id != tokenId) {
+        return response.status(401).send({ error: 'Não autorizado' })
+    }
 
     const userDeletado = await User.findByIdAndDelete(id)
